third_party/blink/web_tests/fast/dom/HTMLImageElement/image-picture-nested-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("canary").clientWidth is 75
PASS currentSrcFileName("canary") is "blue_rect.jpg?0"
PASS document.getElementById("picture_in_picture").clientWidth is 800
PASS currentSrcFileName("picture_in_picture") is "image-set-2x.png?2"
PASS document.getElementById("picture_in_div_in_picture").clientWidth is 800
PASS currentSrcFileName("picture_in_div_in_picture") is "image-set-2x.png?2"
PASS document.getElementById("picture_in_div_with_source_above").clientWidth is 800
PASS currentSrcFileName("picture_in_div_with_source_above") is "image-set-2x.png?2"
PASS document.getElementById("picture_in_the_way").clientWidth is 800
PASS currentSrcFileName("picture_in_the_way") is "image-set-2x.png?2"
PASS document.getElementById("picture_in_picture_1").clientWidth is 800
PASS currentSrcFileName("picture_in_picture_1") is "image-set-2x.png?2"
PASS document.getElementById("picture_in_picture_2").clientWidth is 800
PASS currentSrcFileName("picture_in_picture_2") is "image-set-2x.png?2"
PASS document.getElementById("picture_at_end").clientWidth is 75
PASS currentSrcFileName("picture_at_end") is "blue_rect.jpg?52"
PASS document.getElementById("picture_no_source").clientWidth is 75
PASS currentSrcFileName("picture_no_source") is "blue_rect.jpg?6"
 
    
